+++ src/components/MapViewManager.tsx
@@ -31,8 +31,8 @@ type Props = {
 	rotationEnabled: 1 | 0
 	zoomEnabled: 1 | 0
 	zoomLevel: number;
-	minZoom: number;
-	maxZoom: number;
+	zoomMin: number;
+	zoomMax: number;
 	tilt: number;
 	minTilt: number;
 	maxTilt: number;